pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gds-mnist"
version = "0.1.0"
description = "Geometry-based data selection experiments on MNIST."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "lightning>=2.2.0",
  "matplotlib>=3.8.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pyarrow>=15.0.0",
  "PyYAML>=6.0.0",
  "tqdm>=4.66.0",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
